refactor: move CryptoContext to existing context folder for consistency
Moved app/contexts/CryptoContext.tsx -> app/context/CryptoContext.tsx
to match existing folder structure (AuthContext is in app/context).

Updated imports in:
- app/layout.tsx
- app/messages/page.tsx
- E2EE_IMPLEMENTATION.md
